requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
coverage==4.5.1
dj-database-url==0.5.0
Django==1.11.16
django-crispy-forms==1.7.2
django-tastypie==0.14.2
djangorestframework==3.8.2
Faker==0.9.1
funcsigs==1.0.2
gunicorn==19.9.0
idna==2.7
ipaddress==1.0.22
mixer==6.1.3
more-itertools==4.3.0
pathlib2==2.3.2
Pillow==5.3.0
pluggy==0.8.0
py==1.7.0
pytest==3.9.1
pytest-cov==2.6.0
pytest-django==3.4.3
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.5
requests==2.19.1
scandir==1.9.0
schedule==0.5.0
six==1.11.0
text-unidecode==1.2
urllib3==1.23
whitenoise==4.1
